ask_my_docs/requirements.txt
streamlit
google-generativeai
python-dotenv
faiss-cpu
langchain
langchain_google_genai
PyPDF2
tiktoken >= 0.3.1, < 1.0.0
atlassian-python-api >= 3.41.2, < 4.00
beautifulsoup4 >= 4.12.2, < 5.00
numpy >= 1.24.4, < 2.0.0
regex >= 2023.10.3, < 2024.0.0
pandas >= 2.0.3, < 3.0.0
lxml >= 4.9.3, < 5.0.0
tabulate >= 0.9.0, < 1.0.0
html2text >= 2020.1.16, < 2021.0.0